fix: fix the eval function
It appears the alpha-beta pruning algorithm was not working properly. We
were not checking the output of the `compare_favor` function which meant
we were just checking if the evaluations were not the same. This reduced
the eval function calls on the older example from 716k to 168k.

Also other simple changes.